feature: change design in ConversionTypeCard
